+++ modules/canvaskit/npm_build/types/canvaskit-wasm-tests.ts
@@ -368,9 +368,6 @@ function globalTests(CK: CanvasKit, path?: Path) {
     if (!path) {
         return;
     }
-    const ctx = CK.currentContext();
-    CK.setCurrentContext(ctx);
-    CK.deleteContext(ctx);
     const n = CK.getDecodeCacheLimitBytes();
     const u = CK.getDecodeCacheUsedBytes();
     CK.setDecodeCacheLimitBytes(1000);
@@ -871,12 +868,13 @@ function surfaceTests(CK: CanvasKit, gl?: WebGLRenderingContext) {
     surfaceSeven.delete();
 
     const ctx = CK.GetWebGLContext(canvasEl); // $ExpectType number
+    CK.deleteContext(ctx);
     const grCtx = CK.MakeGrContext(ctx);
-    const surfaceNine = CK.MakeOnScreenGLSurface(grCtx, 100, 400, // $ExpectType Surface
+    const surfaceNine = CK.MakeOnScreenGLSurface(grCtx!, 100, 400, // $ExpectType Surface
         CK.ColorSpace.ADOBE_RGB)!;
 
-    const rt = CK.MakeRenderTarget(grCtx, 100, 200); // $ExpectType Surface | null
-    const rt2 = CK.MakeRenderTarget(grCtx, { // $ExpectType Surface | null
+    const rt = CK.MakeRenderTarget(grCtx!, 100, 200); // $ExpectType Surface | null
+    const rt2 = CK.MakeRenderTarget(grCtx!, { // $ExpectType Surface | null
         width: 79,
         height: 205,
         colorType: CK.ColorType.RGBA_8888,